+++ crates/affinidi-tdk/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - **FEATURE:** Added X25519 key support to `generate_did_key()`
   - `KeyType` also now feature X25519 alongside Ed25519
+- **FEATURE:** Added ability to generate did:peer DIDs directly from TDK using
+  pre-created secrets
+  - This is useful where you already have key material and want to create a did:peer
 
 ## 3rd November 2025 (Release 0.2.6)